fixed lint and build issues
